pyproject.toml
[project]
name = "price-recon"
version = "0.1.0"
description = "Price reconciliation program"
authors = [
    {name = "JasonG",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.11.4,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "ruff (>=0.11.9,<0.12.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]

[tool.poetry]
packages = [{include = "price_recon", from = "src"}]

[tool.ruff]
line-length = 160

[tool.pytest.ini_options]
pythonpath = "."
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
